+++ field_type.go
@@ -36,6 +36,7 @@ const (
 	RECORDS_PER_SECOND   FieldType = 10
 	PID_ARRAY            FieldType = 11
 	DIGEST_ARRAY         FieldType = 12
+	SCAN_MAX_RECORDS     FieldType = 13
 	INDEX_NAME           FieldType = 21
 	INDEX_RANGE          FieldType = 22
 	INDEX_FILTER         FieldType = 23